Move and rename REST API sources (#50)
