+++ src/rest/rest_authentication.c
@@ -19,10 +19,10 @@
 
 #include <string.h>
 
-#include "rest-authentication.h"
-#include "security.h"
-#include "logging.h"
-#include "http_codes.h"
+#include "rest_authentication.h"
+#include "../security.h"
+#include "../logging.h"
+#include "../http_codes.h"
 
 static int validate_authentication_body(json_t *authentication_json)
 {
