+++ src/rest/rest_authentication.h
