+++ src/rest/rest_notifications.c
@@ -19,8 +19,8 @@
 
 #include <string.h>
 
-#include "logging.h"
-#include "punica.h"
+#include "../logging.h"
+#include "../punica.h"
 
 bool valid_callback_url(const char *url)
 {
