+++ src/rest/rest_resources.c
@@ -24,8 +24,8 @@
 #include <sys/syscall.h>
 #include <linux/random.h>
 
-#include "logging.h"
-#include "punica.h"
+#include "../logging.h"
+#include "../punica.h"
 
 typedef struct
 {
