+++ src/rest/rest_utils.c
@@ -19,9 +19,9 @@
 #include <string.h>
 #include <uuid/uuid.h>
 
-#include "rest-utils.h"
+#include "rest_utils.h"
 
-#include "punica.h"
+#include "../punica.h"
 
 #define DATABASE_UUID_KEY_BIT       0x1
 #define DATABASE_PSK_KEY_BIT        0x2
