+++ src/settings.c
@@ -25,7 +25,7 @@
 #include "settings.h"
 #include "version.h"
 #include "security.h"
-#include "rest-core-types.h"
+#include "rest/rest_core_types.h"
 
 #define DATABASE_UUID_KEY_BIT       0x1
 #define DATABASE_PSK_KEY_BIT        0x2
